requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jax==0.4.38
jaxlib==0.4.38
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
moviepy==2.1.2
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==10.4.0
proglog==0.1.10
protobuf>=4.21.0
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.37
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
tensorflow==2.18.0
tensorflow_intel==2.13.0
keras==3.8.0
